Removing special re-focusing for focused items
Focused items should automatically be kept in memory by activation. The initial focus is just to help attention, but never really allows the conversation to move away from it as it is repeatedly refocused after that. We may end up doing the opposite and banning re-focus for a period of time after an item leaves focus.